server/requirements.txt
openai
uvicorn
openai-whisper
fastapi
starlette
pydantic